refactor : 임시폴저에 저장된 파일들을 정상화
에러핸들링 관련해서 에러가 발생해서 그부분은 롤백함
