+++ client/src/presentation/components/snug/chat-input-box/index.tsx
@@ -1,9 +1,11 @@
-import React from "react";
+import React, { useState } from "react";
 import styled from "styled-components";
 import ClipWhite from "assets/clip-white.png";
 import AtWhite from "assets/at-white.png";
 import FaceWhite from "assets/face-white.png";
 import { IconBox } from "presentation/components/atomic-reusable/icon-box";
+import { useMessagesDispatch } from "contexts/messages-context";
+import dubu from "assets/dubu.png";
 
 const InputWrapper = styled.section`
   width: 100%;
@@ -46,14 +48,48 @@ const StyledInput = styled.input.attrs({
     outline: none;
   }
 `;
-
 export const ChatInputBox: React.FC = () => {
+  const KEY_PRESS_EVENT_KEY = "Enter";
+  const [message, setMessage] = useState("");
+  const [id, setId] = useState(0);
+  const dispatch = useMessagesDispatch();
+
+  const inputChangeEventHandler = (
+    event: React.ChangeEvent<HTMLInputElement>
+  ) => {
+    setMessage(event.target.value);
+  };
+
+  //이 부분은 mock 데이터로 되어 있으니 차후 수정이 필요함
+  const inputKeyPressEventHandler = (
+    event: React.KeyboardEvent<HTMLInputElement>
+  ) => {
+    event.stopPropagation();
+    if (event.key !== KEY_PRESS_EVENT_KEY) return;
+    if (!message.trim()) return;
+    if (!dispatch) return;
+    dispatch({
+      type: "CREATE",
+      id: id + 1,
+      name: "두부",
+      imageSrc: dubu,
+      timestamp: new Date().toLocaleTimeString(),
+      contents: message
+    });
+    setId(id + 1);
+    setMessage("");
+  };
+
   return (
     <InputWrapper>
       <MarginBox></MarginBox>
       <CustomInput>
         <IconBox imageSrc={ClipWhite}></IconBox>
-        <StyledInput></StyledInput>
+        <StyledInput
+          value={message}
+          onChange={inputChangeEventHandler}
+          onKeyPress={inputKeyPressEventHandler}
+        ></StyledInput>
         <IconBox imageSrc={AtWhite}></IconBox>
         <IconBox imageSrc={FaceWhite}></IconBox>
       </CustomInput>
