+++ client/src/presentation/components/snug/message-card/index.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import { MessageCardImage } from "presentation/components/atomic-reusable/message-card-image";
 import { MessageCardContents } from "./message-card-contents";
+import { Message } from "core/entity/message";
 
 const MessageCardBox = styled.section`
   width: 100%;
@@ -13,29 +14,22 @@ const MarginBox = styled.section`
   width: 10px;
 `;
 
-interface PropTypes {
-  imageSrc: string;
-  userName: string;
-  timestamp: string;
-  message: string;
-}
-
-export const MessageCard: React.FC<PropTypes> = ({
+export const MessageCard: React.FC<Message> = ({
   imageSrc,
-  userName,
+  name,
   timestamp,
-  message
+  contents
 }) => {
   return (
     <MessageCardBox>
-      <MarginBox></MarginBox>
+      <MarginBox />
       <MessageCardImage imageSrc={imageSrc}></MessageCardImage>
       <MessageCardContents
-        userName={userName}
+        userName={name}
         timestamp={timestamp}
-        message={message}
+        message={contents}
       ></MessageCardContents>
-      <MarginBox></MarginBox>
+      <MarginBox />
     </MessageCardBox>
   );
 };
