+++ client/src/presentation/pages/snug/message-section/content.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import { ChatContent } from "presentation/components/snug/chat-container";
 import { ChatInputBox } from "presentation/components/snug/chat-input-box";
+import { MessageContextProvider } from "contexts/messages-context";
 
 const MessageSectionContentWrapper = styled.section`
   width: 100%;
@@ -13,9 +14,11 @@ const MessageSectionContentWrapper = styled.section`
 
 export const MessageSectionContent: React.FC = () => {
   return (
-    <MessageSectionContentWrapper>
-      <ChatContent></ChatContent>
-      <ChatInputBox></ChatInputBox>
-    </MessageSectionContentWrapper>
+    <MessageContextProvider>
+      <MessageSectionContentWrapper>
+        <ChatContent />
+        <ChatInputBox />
+      </MessageSectionContentWrapper>
+    </MessageContextProvider>
   );
 };
